requirements.txt
flask==3.0.2
line-bot-sdk>=3.0.0
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
requests==2.31.0
python-dotenv==1.0.1
openai==0.28.1
Pillow>=9.0.0
gunicorn==21.2.0 